+++ src/main/java/com/learn/spring/model/Book.java
@@ -1,0 +1,9 @@
+package com.learn.spring.model;
+
+import lombok.Data;
+
+@Data
+public class Book {
+
+	private String author;
+}