+++ src/main/java/com/learn/spring/model/Inventory.java
@@ -1,4 +1,4 @@
-package com.learn.spring;
+package com.learn.spring.model;
 
 import lombok.Data;
 
@@ -16,5 +16,13 @@ public class Inventory {
 	private String name;
 
 	private Double price;
-
+	
+	private String remarks;
+	
+	private Book bookDetails;
+	
+	private String registration;
+	
+	private String address;
+	
 }