+++ changeLog.md
@@ -2,3 +2,13 @@ Commit #2 on 04-03-21
 ---------------------
 - Bean scope
 - Bean life cycle
+
+Commit #3 on 05-03-21
+---------------------
+- Bean Post Processor
+- Bean Definition Inheritance
+- Injecting Inner Beans
+
+Commit #4 on 05-03-21
+---------------------
+- Injecting Collection